refactor : SecureRandom 단위 테스트 작성
